temp commit

+++ src/curfil/random_tree_image_gpu.cu
@@ -1205,12 +1205,15 @@ TreeNodeData getTreeNode(const int nodeNr, const boost::shared_ptr<const TreeNod
     return data;
 
 }
-
 __global__ void classifyKernel(
         float* output, int tree,
         const int16_t imageWidth, const int16_t imageHeight,
-        const LabelType numLabels, bool useDepthImages, size_t* nodeOffsets ) {
-
+       const LabelType numLabels, bool useDepthImages, unsigned int* nodeOffsets ) {
+//__global__ void classifyKernel(
+  //      float* output, int tree,
+ //       const int16_t imageWidth, const int16_t imageHeight,
+ //       const LabelType numLabels, bool useDepthImages ) {
+	//size_t* nodeOffsets;
     const unsigned int x = blockIdx.x * blockDim.x + threadIdx.x;
     if (x >= imageWidth) {
         return;
@@ -1400,8 +1403,9 @@ void determineMaxProbabilities(const cuv::ndarray<float, cuv::dev_memory_space>&
 }
 
 void classifyImage(int treeCacheSize, cuv::ndarray<float, cuv::dev_memory_space>& output, const RGBDImage& image,
-        LabelType numLabels, const boost::shared_ptr<const TreeNodes>& treeData, bool useDepthImages,cuv::ndarray<size_t, cuv::dev_memory_space>& nodeOffsets) {
-
+        LabelType numLabels, const boost::shared_ptr<const TreeNodes>& treeData, bool useDepthImages,cuv::ndarray<unsigned int, cuv::dev_memory_space>& nodeOffsets) {
+//void classifyImage(int treeCacheSize, cuv::ndarray<float, cuv::dev_memory_space>& output, const RGBDImage& image,
+	//        LabelType numLabels, const boost::shared_ptr<const TreeNodes>& treeData, bool useDepthImages) {
     std::set<const RGBDImage*> images;
     images.insert(&image);
 
@@ -1433,9 +1437,13 @@ void classifyImage(int treeCacheSize, cuv::ndarray<float, cuv::dev_memory_space>
 
     cudaSafeCall(cudaFuncSetCacheConfig(classifyKernel, cudaFuncCachePreferL1));
 
-    classifyKernel<<<blockSize, threads, 0, stream>>>(output.ptr(), tree,
-            image.getWidth(), image.getHeight(),
-            numLabels, useDepthImages, nodeOffsets.ptr());
+   classifyKernel<<<blockSize, threads, 0, stream>>>(output.ptr(), tree,
+           image.getWidth(), image.getHeight(),
+           numLabels, useDepthImages, nodeOffsets.ptr());
+
+  //  classifyKernel<<<blockSize, threads, 0, stream>>>(output.ptr(), tree,
+  //          image.getWidth(), image.getHeight(),
+  //          numLabels, useDepthImages);
 
     cudaSafeCall(cudaStreamSynchronize(stream));
 }
